# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "video2ctx API",
    "version": "1.0.0",
    "license": {
      "name": "Apache License 2.0",
      "identifier": "Apache-2.0"
    },
    "description": "Consumer-facing contract for the hosted video2ctx API. First-party application and operator routes are documented separately at https://docs.video2ctx.dev/internals/overview."
  },
  "security": [],
  "servers": [
    {
      "url": "https://api.video2ctx.dev",
      "description": "Hosted video2ctx API"
    }
  ],
...
  "paths": {
...
    "/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "listProjects",
        "summary": "List projects",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "responses": {
          "200": {
            "description": "The user’s projects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "projects"
                  ],
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-video2ctx-audience": "consumer"
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "createProject",
        "summary": "Create a project",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreated"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
    "/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "getProject",
        "summary": "Get a project and its saved items",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource identifier.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-video2ctx-audience": "consumer"
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "deleteProject",
        "summary": "Delete a project",
        "description": "Deletes the project’s private R2 objects and AI Search items before cascading its D1 metadata.",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource identifier.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
    "/v1/projects/{id}/items": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "addProjectItem",
        "summary": "Save material to a project",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource identifier.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project item created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
    "/v1/imports": {
      "post": {
        "tags": [
          "Research"
        ],
        "operationId": "createImport",
        "summary": "Start a durable import job",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Optional caller-provided idempotency key.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImportRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Import queued or an existing idempotent job returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "Research"
        ],
        "operationId": "getJob",
        "summary": "Get import job status",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource identifier.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Durable job status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
...
    "/v1/projects/{id}/exports": {
      "post": {
        "tags": [
          "Exports"
        ],
        "operationId": "createProjectExport",
        "summary": "Create a project export",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource identifier.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Export created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Export"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
    "/v1/exports/{id}/download": {
      "get": {
        "tags": [
          "Exports"
        ],
        "operationId": "downloadExport",
        "summary": "Download an export",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Resource identifier.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export file.",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
...
    "/v1/notifications": {
      "get": {
        "tags": [
          "Monitoring"
        ],
        "operationId": "listNotifications",
        "summary": "List recent notifications",
        "security": [
          {
            "bearerApiKey": []
          },
          {
            "sessionCookie": []
          },
          {
            "cliSession": []
          },
          {
            "apiKey": []
          },
          {
            "demoUser": []
          }
        ],
        "responses": {
          "200": {
            "description": "Up to 100 recent notifications.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "notifications"
                  ],
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "x-video2ctx-audience": "consumer"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Better Auth session cookie. A production deployment may use the secure-prefixed cookie name; same-origin browser requests send it automatically."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Legacy-compatible header for a permanent personal API key. Prefer Authorization: Bearer aty_…."
      },
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Permanent personal API key created in the dashboard. Example: curl -H \"Authorization: Bearer aty_…\" https://your-host/v1/projects"
      },
      "cliSession": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "CLI session",
        "description": "Short-lived, revocable session issued by the video2ctx device authorization flow. The official CLI stores and sends this value; never paste it into prompts or logs."
      },
      "demoUser": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Demo-User",
        "description": "Local development only. Any stable value selects an isolated demo account. Rejected in production."
      },
      "stripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe-generated signature over the exact raw request body."
      }
    },
...
    "schemas": {
...
      "Project": {
        "allOf": [
          {
            "type": "object",
            "description": "A persisted D1 record. Database timestamps are Unix milliseconds.",
            "additionalProperties": true
          },
          {
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "item_count": {
                "type": "integer"
              }
            }
          }
        ]
      },
...
      "Export": {
        "type": "object",
        "description": "A persisted D1 record. Database timestamps are Unix milliseconds.",
        "additionalProperties": true
      },
...
    }
  }
}